llm-service/pyproject.toml
[project]
name = "llm-service"
version = "0.1.0"
description = "LLM Service with OpenAI-compatible API"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "llama-index-core>=0.10.0",
    "llama-index-llms-huggingface>=0.2.0",
    "torch>=2.1.0",
    "accelerate>=0.24.0",
    "pydantic>=2.5.0",
    "loguru>=0.7.0",
    "pydantic-settings>=2.1.0",
    "llama-index-llms-openai-like",
]
